requirements.txt
Adafruit-BBIO==1.1.1
Adafruit-Blinka==5.8.0
adafruit-circuitpython-busdevice==5.0.1
adafruit-circuitpython-vl6180x==1.2.3
adafruit-extended-bus==1.0.0
Adafruit-PlatformDetect==2.22.1
Adafruit-PureIO==1.1.7
aenum==2.2.4
Click==7.0
colorama==0.3.7
decorator==4.3.0
distro==1.3.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.2
pyctrl==0.4.3
pyftdi==0.52.0
pyserial==3.5b0
python-can==3.3.4
pyuarm==2.2.5.3
pyuf==1.3
pyusb==1.1.0
rcpy==0.5.1
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
scipy==1.1.0
sysv-ipc==1.0.1
Werkzeug==0.14.1
wrapt==1.12.1